security-control-py/requirements.txt
paho-mqtt==2.1.0
redis==5.1.1
fastapi==0.115.0
uvicorn==0.30.6
serial
